pyproject.toml
[project]
name = "busqueda-laptop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "gradio>=6.6.0",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.2",
]
